requirements.txt
PyQt5
pyinstaller
pillow
haralyzer